add upload file

+++ src/main/java/com/example/tiktok/repositories/UserRepository.java
@@ -1,4 +1,5 @@
 package com.example.tiktok.repositories;
+
 import com.example.tiktok.entities.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
@@ -6,6 +7,6 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Integer> {
+public interface UserRepository extends JpaRepository<User, Long> {
     User findByEmail(String email);
 }
